add change pp - header CSS

+++ JS/firebase_profil.js
@@ -18,20 +18,24 @@ const supprimer_le_compte = document.getElementById('supp')
 
 supprimer_le_compte.addEventListener('click', e =>{
     const user = firebase.auth().currentUser
-  // fonction delete
-  user.delete().then(function(){
-      document.location.pathname='index.html'
-  }).catch(function(error){
-    console.log("erreur de suppression")
-  })
+    // fonction delete
+    if (confirm("Etes vous sur de vouloir supprimer votre compte ?")) {
+        user.delete().then(function(){
+            document.location.pathname='index.html'
+        }).catch(function(error){
+            console.log("erreur de suppression")
+        })
+    }
 })
 
 //deconnexion
 const deconnexion = document.getElementById('deco')
 
 deconnexion.addEventListener('click', e =>{
-    firebase.auth().signOut();
-    document.location.pathname='index.html'
+    if (confirm("Déconnecter ?")) {
+        firebase.auth().signOut();
+        document.location.pathname='index.html'
+    }
 })
 
 firebase.auth().onAuthStateChanged(firebaseUser =>{
@@ -52,23 +56,37 @@ const nom = document.getElementById('nom')
 const prenom = document.getElementById('prenom')
 const email = document.getElementById('email')
 const tel = document.getElementById('tel')
+const lateAbs = document.getElementById('lateAbs')
+const pp = document.getElementById('photo_prof')
+
 
 
 const fb = firebase.firestore();
 
 window.onload = () => {
     setTimeout(() => {
-        fb.collection('Personnes_connectés').doc(firebase.auth().currentUser.uid).get().then(doc => {
+        fb.collection('Personnes_connectés').doc(firebase.auth().currentUser.uid).onSnapshot(doc => {
+            console.log(doc.data().pp)
+            if (doc.data().pp != undefined && doc.data().pp != "") {
+                pp.src = doc.data().pp
+            }
             nom.innerHTML = doc.data().Nom
             prenom.innerHTML = doc.data().Prenom
             email.innerHTML = doc.data().Email
-            telephone.innerHTML = doc.data().Telephone
+            tel.innerHTML = doc.data().Telephone
+            if(doc.data().autorisation == 3){
+                console.log(lateAbs)
+                lateAbs.innerHTML = `<h3>Nombre de retards</h3>
+                <p>${doc.data().late}</p>
+                <h3>Nombre d'absences</h3>
+                <p>${doc.data().absence}</p>`
+            }
         })
-    },700)
+    },900)
 }
 
-var uploader = document.getElementById('uploader');
-var fileButton = document.getElementById('fileButton');
+
+const fileButton = document.getElementById('fileButton');
 
 //Listen for file selection
 
@@ -77,22 +95,21 @@ fileButton.addEventListener('change', function(e){
     var file = e.target.files[0]
 
     //create storage ref
-    var storageRef = firebase.storage().ref("toaster/" + file.name)
+    var storageRef = firebase.storage().ref("IdPicures/" + file.name)
 
     //upload file
      var task = storageRef.put(file);
     console.log("hello")
-    //update progress bar
     task.on('state_changed',
         function progress (snapshot){
-            var percentage = (snapshot.bytesTransferred / snapshot.totalBytes)*100;
-            uploader.value = percentage;
         },
         function error(err){
-
+            console.log(err)
         },
         function complete (){
-
+            task.snapshot.ref.getDownloadURL().then(url => {
+                fb.collection('Personnes_connectés').doc(firebase.auth().currentUser.uid).update({"pp": url})
+            })
         }
     );
